Defining Animation
The first animation-specific function that we use is trigger(). This allows us to define the name of the animation in the first parameter, and then an array of other animation-specific functions in the next.
We've made a new animation called goals here. Next, we use transition() to define a transition between two different states. States can be defined explicitely through the state function, or we can use a wildcard * as a state.

There are a variety of transition states, but the one we're going to use is * => *:
Next, we're going to use the query function to target DOM elements that are only entering the DOM
Next, we're creating a stagger animation with a delay of 300ms between each element.

After that, we're animating the elements with a .6s ease-in animation, and using keyframes to allow us to create a multi-step animation for each item.
finally Applying the Animation in the Template